search_api/requirements.txt
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.37
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
